+++ src/Symfony/Component/Config/CHANGELOG.md
@@ -6,6 +6,12 @@ CHANGELOG
 
  * removed `ReferenceDumper` class
 
+2.7.0
+-----
+
+ * added `ConfigCacheInterface`, `ConfigCacheFactoryInterface` and a basic `ConfigCacheFactory`
+   implementation to delegate creation of ConfigCache instances
+   
 2.2.0
 -----